Keep ltype integer for most frequent impute tpye in target imputer

+++ evalml/tests/integration_tests/test_data_checks_and_actions_integration.py
@@ -222,7 +222,7 @@ def test_data_checks_impute_cols(problem_type):
         expected_pipeline_class = BinaryClassificationPipeline
         y_expected = ww.init_series(
             pd.Series([0, 1, 1, 1, 1]),
-            logical_type="Double",
+            logical_type="Integer",
         )
 
     elif problem_type == "multiclass":
@@ -231,7 +231,7 @@ def test_data_checks_impute_cols(problem_type):
         expected_pipeline_class = MulticlassClassificationPipeline
         y_expected = ww.init_series(
             pd.Series([0, 1, 2, 2, 2]),
-            logical_type="Double",
+            logical_type="Integer",
         )
 
     else:
